Fixed colliders and refactored dash
