+++ Assets/Scripts/DashScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5be18d89830954885929a7b5ba6d15a7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
